feat: Implement individual checklist app for readability with CRUD functionality and JSON persistence
